Feat : Confirm을 활용한 체크

+++ src/main/webapp/jsp/usr/article/list.jsp
@@ -20,7 +20,7 @@
                 <a class="w-[40px] hover:underline hover:text-[red]" href="/usr/article/detail/<%=article.getId()%>"><%=article.getId()%></a>
                 <!-- flex-grow : 성장성 1 -->
                 <a class="flex-grow hover:underline hover:text-[red]" href="/usr/article/detail/<%=article.getId()%>"><%=article.getTitle()%></a>
-                <a class="w-[100px] hover:underline hover:text-[red]" href="/usr/article/delete/<%=article.getId()%>">삭제</a>
+                <a onclick="if ( !confirm('정말로 삭제하시겠습니까?') ) return false;" class="w-[100px] hover:underline hover:text-[red]" href="/usr/article/delete/<%=article.getId()%>">삭제</a>
             </li>
             <% } %>
         </ul>
